# openapi/openapi.yaml
openapi: 3.0.0
info:
  title: VoyageStack API
  version: 1.0.0
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
  schemas:
    Expense:
      type: object
      properties:
        id: { type: string, format: uuid }
        tripId: { type: string, format: uuid }
        payerId: { type: string, format: uuid }
        amount: { type: number, format: float, description: "Total expense amount" }
        currency: { type: string, example: "TWD" }
        description: { type: string, nullable: true }
        splitInfo:
          type: object
          additionalProperties: { type: number }
          description: "Map of userId -> share amount owed by that user (sum should equal amount)"
        createdAt: { type: string, format: date-time }
...
    Task:
      type: object
      properties:
        id: { type: string, format: uuid }
        tripId: { type: string, format: uuid }
        assignedUserId: { type: string, format: uuid, nullable: true }
        title: { type: string }
        category:
          type: string
          enum: [general, esim, visa, accommodation, transport]
        status:
          type: string
          enum: [todo, in_progress, done]
        dueDate: { type: string, format: date, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    ItineraryItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tripId:
          type: string
          format: uuid
        day:
          type: integer
          nullable: true
        title:
          type: string
        category:
          type: string
          enum: [food, lodging, attraction, activity, transport, admin]
          nullable: true
          description: Spot category for classification
        coverImage:
          type: string
          nullable: true
          description: Cover image URL (populated from Instagram OG metadata)
        lat:
          type: number
          nullable: true
        lng:
          type: number
          nullable: true
        order:
          type: integer
        sourceUrl:
          type: string
          nullable: true
        note:
          type: string
          nullable: true
security:
  - BearerAuth: []
paths:
...
  /api/trips:
    get:
      summary: Get my trips
      responses:
        '200':
          description: List of trips
    post:
      summary: Create a trip
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title]
              properties:
                title:
                  type: string
                startDate:
                  type: string
                  format: date
                endDate:
                  type: string
                  format: date
      responses:
        '201':
          description: Created trip
  /api/trips/join:
    post:
      summary: Join trip by invite code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [inviteCode]
              properties:
                inviteCode:
                  type: string
      responses:
        '200':
          description: Joined trip
  /api/trips/{tripId}:
    get:
      summary: Get trip by ID
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Trip detail
        '403':
          description: Forbidden
        '404':
          description: Not found
    patch:
      summary: Update trip
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Updated trip
        '403':
          description: Forbidden
  /api/trips/{tripId}/itinerary:
    get:
      summary: Get trip itinerary items
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: string
        - in: query
          name: bucket
          required: false
          schema:
            type: boolean
          description: If true, returns only bucket items (day IS NULL)
      responses:
        '200':
          description: List of itinerary items
    post:
      summary: Add itinerary item
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [day, title]
              properties:
                day:
                  type: integer
                title:
                  type: string
                category:
                  type: string
                  enum: [food, lodging, attraction, activity, transport, admin]
                  description: Spot category for classification
                coverImage:
                  type: string
                  description: Cover image URL (populated from Instagram OG metadata)
                lat:
                  type: number
                lng:
                  type: number
                sourceUrl:
                  type: string
                note:
                  type: string
      responses:
        '201':
          description: Created item
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/ItineraryItem'
                  error:
                    nullable: true
...
  /api/trips/{tripId}/expenses:
    get:
      summary: List expenses for a trip
      tags: [Expenses]
      parameters:
        - { name: tripId, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: List of expenses (sorted by createdAt DESC)
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Expense' }
                  error: { nullable: true }
    post:
      summary: Create an expense
      tags: [Expenses]
      parameters:
        - { name: tripId, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [payerId, amount, splitInfo]
              properties:
                payerId: { type: string, format: uuid }
                amount: { type: number, minimum: 0.01 }
                currency: { type: string, default: "TWD" }
                description: { type: string }
                splitInfo:
                  type: object
                  additionalProperties: { type: number }
                  description: "Map of userId -> share amount owed (sum should equal amount)"
      responses:
        '201':
          description: Created expense
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/Expense' }
                  error: { nullable: true }
        '400': { description: Validation error (INVALID_PAYER, INVALID_AMOUNT, INVALID_SPLIT) }
        '403': { description: Forbidden — Owner or Editor role required }
...
  /api/trips/{tripId}/tasks:
    get:
      summary: List tasks for a trip
      tags: [Tasks]
      description: Sorted by dueDate ASC NULLS LAST, then status, then createdAt ASC.
      parameters:
        - { name: tripId, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: List of tasks
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Task' }
                  error: { nullable: true }
    post:
      summary: Create a task
      tags: [Tasks]
      parameters:
        - { name: tripId, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title]
              properties:
                title: { type: string }
                category:
                  type: string
                  enum: [general, esim, visa, accommodation, transport]
                  default: general
                status:
                  type: string
                  enum: [todo, in_progress, done]
                  default: todo
                assignedUserId: { type: string, format: uuid }
                dueDate: { type: string, format: date, nullable: true }
      responses:
        '201':
          description: Created task
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/Task' }
                  error: { nullable: true }
        '400': { description: Validation error (INVALID_TITLE, INVALID_CATEGORY, INVALID_STATUS) }
        '403': { description: Forbidden — Owner or Editor role required }
...